:headstone: Abandoning C++ Project
I've made the decision to put this project to bed now. I'm extremely pleased with how much I implemented, but in the end, my architecture for the game fell short. My new ideas for structuring the game's code are proving to be arduous to implement in C++, and I no longer have enough free time to justify struggling with it.

The good news is that I'm hoping to begin a new project working on the same idea but using TypeScript instead. This will make development rapid and less painstaking, and with it being a web application, it will lend itself well to multiplayer gameplay across a multitude of devices. I will hopefully spend a lot more time coding the game itself instead of all the systems that underpin it.